Added the calculation of offset in range of motor from center of screen (-1 to 1). Also added a system to send over the network the values.
There is a place on the network for [mid-left_x, top_x, mid-right_x, bottom_x]
[mid-left_y, top_y, mid-right_y, bottom_y]
[mid-left_exists, top_exists, mid-right_exists, bottom_exists]

The *_x is the offset between -1 and 1 for the x (negative is the robot needs to look left more)
the *_y is the offset between -1 and 1 for the y (negative is the robot needs to look up more)
the *_exists is a boolean that is true if we know the location of that target

The boolean is important because the *_x and *_y will be set the the previous location the target was found even if it is no longer found. In other words, don't use the the values in *_x or *_y if *_exists is false.